Split new PAL types into new files
